Use a changelog and issues instead of a Todo file.
I've moved all of the 'ToDo' items out of this file into GitHub issues.  I want to get a new developer on this to this project and I think it is better if the tasks are linked to issues (so that they can be resolved by PRs) and there is a place for additional discussion relating to issues. I just did a straight copy/paste of the todo's in this file to issues so modifications to issues are welcome.

I am proposing this file really should be a changelog for the project.  See [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) for additional documentation.